feat(elsa): refactor main
- move main to cmd
